new exception method introduced

+++ backend/src/main/java/com/cs319group3/backend/Services/ServiceImpls/TaskTypeServiceImpl.java
@@ -1,12 +1,18 @@
 package com.cs319group3.backend.Services.ServiceImpls;
 
+import com.cs319group3.backend.Components.ApiResponse;
 import com.cs319group3.backend.DTOMappers.TaskTypeMapper;
 import com.cs319group3.backend.DTOs.TaskTypeDTO;
+import com.cs319group3.backend.Entities.RequestEntities.TAWorkloadRequest;
 import com.cs319group3.backend.Entities.TaskType;
 import com.cs319group3.backend.Repositories.CourseRepo;
+import com.cs319group3.backend.Repositories.TAWorkloadRequestRepo;
 import com.cs319group3.backend.Repositories.TaskTypeRepo;
 import com.cs319group3.backend.Services.TaskTypeService;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -15,10 +21,13 @@ import java.util.List;
 public class TaskTypeServiceImpl implements TaskTypeService {
 
     @Autowired
-    CourseRepo courseRepo;
+    private CourseRepo courseRepo;
 
     @Autowired
-    TaskTypeRepo taskTypeRepo;
+    private TaskTypeRepo taskTypeRepo;
+
+    @Autowired
+    private TAWorkloadRequestRepo taWorkloadRequestRepo;
 
     @Override
     public boolean createTaskType(TaskTypeDTO dto, int courseId) {
@@ -32,11 +41,18 @@ public class TaskTypeServiceImpl implements TaskTypeService {
     }
 
     @Override
-    public boolean deleteTaskType(int courseId, String taskTypeName) {
+    public ResponseEntity<?> deleteTaskType(int courseId, String taskTypeName) {
+
+
+        List<TAWorkloadRequest> requests = taWorkloadRequestRepo.findByTaskType_TaskTypeName(taskTypeName);
+        if(!requests.isEmpty()) {
+            return ResponseEntity.badRequest().body(new ApiResponse(false, "Task type " + taskTypeName + " already exists"));
+        }
+
         if(taskTypeRepo.findByTaskTypeNameAndCourse_CourseId(taskTypeName, courseId).isPresent()){
             TaskType taskType = taskTypeRepo.findByTaskTypeNameAndCourse_CourseId(taskTypeName, courseId).get();
             taskTypeRepo.delete(taskType);
-            return true;
+            return new ResponseEntity<>(true, HttpStatus.ACCEPTED);
         }
         else{
             throw new RuntimeException("Task type with name " + taskTypeName + " not found.");
